Profile search: location autocomplete
- Add location autocomplete to profile search page

#3570
