+++ assets/js/backbone/app.js
@@ -12,8 +12,8 @@ require('jquery-ui-dist/jquery-ui');
 var _ = require('underscore');
 var Backbone = require('backbone');
 
-require('./global-utils');
-require('./global-search-utils');
+// Import global functions
+require('./globals');
 
 Backbone.$ = $;
