+++ assets/js/backbone/globals/index.js
@@ -1,0 +1,5 @@
+require('./backbone-utils');
+require('./utils');
+require('./autocomplete-utils');
+require('./search-utils');
+require('./validation-utils');
