+++ config/application.js
@@ -12,6 +12,8 @@ module.exports = {
   hostName: process.env.HOST || 'localhost:3000',
 
   usajobsURL: process.env.USAJOBS_URL || 'https://www.usajobs.gov',
+
+  usajobsDataURL: process.env.USAJOBS_DATA_URL || 'https://data.usajobs.gov',
   
   agencyportalURL: process.env.AGENCYPORTAL_URL || 'https://agencyportal.usajobs.gov',
   // redirect domain not matching hostName
