+++ environment.json
@@ -35,6 +35,7 @@
     "SYSTEM_NAME": "Open Opportunities",
     "TASK_STATE": "draft",
     "USAJOBS_URL": "https://www.usajobs.gov",
+    "USAJOBS_DATA_URL": "",
     "AGENCYPORTAL_URL": "https://localhost:44373",
     "VALIDATE_DOMAINS": true
 }
