+++ views/layout.ejs
@@ -13,6 +13,7 @@
     draftAdminOnly = <%- JSON.parse(draftAdminOnly) %>;
     version = <%- JSON.stringify(version) %>;
     usajobsURL = <%- JSON.stringify(usajobsURL) %>;
+    usajobsDataURL = <%- JSON.stringify(usajobsDataURL) %>;
     </script>
     <script defer src="/js/fontawesome-all.js"></script>
